elf: Remove -fno-tree-loop-distribute-patterns usage on dl-support
Besides the option being gcc specific, his approach is still fragile
since we don't how if this will be the only optimization option gcc
will make loops to memset.

Instead, it adds a __memset_generic internal symbol on all ports to
avoid the iFUNC plt and a symbol redirection to header to include on
TU that required it.